refactor(assistant): extract AssistantController
