+++ frontend/src/hooks/conversation/useConversationEngine.js
@@ -1,4 +1,4 @@
-// frontend\src\hooks\assistant\useConversationEngine.js
+// frontend\src\hooks\conversation\useConversationEngine.js
 import { useState, useCallback } from "react";
 
 export default function useConversationEngine({ flow }) {
